Add files via upload
Poker hand game. IntelliJ takes a deck of cards (52 cards) and takes a set of  5 cards known as community cards. After this it distributes two cards to each of the two players. 
Then it determines the best possible combination by checking different combinations of each hand by determining the worth of each cards in a hand. After finalizing the best possible combination for each hand, the IntelliJ determines whether player 1 or player 2 is winner based on which of them has a hand worth more.